requirements.txt
Django
Pillow
huggingface_hub
django-cors-headers